Revert "slippery fix"
This reverts commit 450720ec91f93981e3ba2903ce352c96c39f7b47.

+++ Assets/Scripts/playerMovement.cs
@@ -57,11 +57,12 @@ public class playerMovement : MonoBehaviour
        
 
         //Get input axes
-        int x = (int)Input.GetAxis("Horizontal");
-        int y = (int)Input.GetAxis("Vertical");
+        float x = Input.GetAxis("Horizontal");
+        float y = Input.GetAxis("Vertical");
+
 
         //Move with local dir
-
+       
         move = new Vector3(x, 0f, y).normalized;
